requirements.txt
nltk==3.8.1
pandas==2.1.1
langdetect==1.0.9
translate==3.6.1
scikit-learn==1.3.1
gensim==4.3.2
numpy==1.26.0
pyspellchecker==0.7.2
openpyxl==3.1.2
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0